Fix options page

+++ backgroundscripts/background.js
@@ -1,5 +1,4 @@
 chrome.browserAction.onClicked.addListener(function(activeTab) {
-  console.log(activeTab);
   chrome.tabs.executeScript({
     file: 'assets/contentscript.js',
     allFrames : true,
@@ -16,30 +15,32 @@ chrome.browserAction.onClicked.addListener(function(activeTab) {
   });
 });
 
-chrome.runtime.onInstalled.addListener(() => {
-  chrome.storage.local.clear();
-  chrome.storage.local.set({
-    regexList: {
-      type0: true,
-      type1: true,
-      type2: true,
-      type3: true,
-      type4: true,
-      type5: true,
-    },
-    outputFormat: {
-      hour: true,
-      minute: true,
-      month: true,
-      day: true,
-      highlight: true,
-      validFor100Years: true,
-    },
-    alertBoxes: {
-      info: true,
-      whatsnew: true,
-    },
-  }, () => {
-    chrome.runtime.openOptionsPage();
-  });
+chrome.runtime.onInstalled.addListener((details) => {
+  if(details.reason === 'install') {
+    chrome.storage.local.clear();
+    chrome.storage.local.set({
+      regexList: {
+        type0: true,
+        type1: true,
+        type2: true,
+        type3: true,
+        type4: true,
+        type5: true,
+      },
+      outputFormat: {
+        hour: true,
+        minute: true,
+        month: true,
+        day: true,
+        highlight: true,
+        validFor100Years: true,
+      },
+      alertBoxes: {
+        info: true,
+        whatsnew: true,
+      },
+    }, () => {
+      chrome.runtime.openOptionsPage();
+    });
+  }
 });
